+++ src/features/usage-consolidation-accordion/summary-table/summary-table.js
@@ -21,6 +21,8 @@ import {
   entityTypes,
 } from '../../../constants';
 
+import style from './summary-table.css';
+
 const propTypes = {
   contentData: PropTypes.array,
   costPerUseData: costPerUseShape.CostPerUseReduxStateShape.isRequired,
@@ -28,18 +30,20 @@ const propTypes = {
   customProperties: PropTypes.object,
   entityType: PropTypes.string.isRequired,
   id: PropTypes.string.isRequired,
+  isExportDisabled: PropTypes.bool,
+  onExportTitles: PropTypes.func,
   onViewTitles: PropTypes.func,
-  year: PropTypes.number.isRequired,
 };
 
 const SummaryTable = ({
   costPerUseData,
   customProperties,
   id,
-  year,
   costPerUseType,
   onViewTitles,
   entityType,
+  onExportTitles,
+  isExportDisabled,
   ...rest
 }) => {
   const intl = useIntl();
@@ -99,19 +103,30 @@ const SummaryTable = ({
               </Button>
             ) : null
             }
-            <Button
-              buttonStyle="dropdownItem fullWidth"
-              role="menuitem"
-              onClick={onToggle}
-              marginBottom0
-            >
-              <Icon
-                icon="download"
-                size="small"
+            <div>
+              <Button
+                buttonStyle="dropdownItem fullWidth"
+                role="menuitem"
+                onClick={() => {
+                  onExportTitles(true);
+                  onToggle();
+                }}
+                disabled={isExportDisabled}
+                marginBottom0
               >
-                <FormattedMessage id="ui-eholdings.usageConsolidation.summary.actions.export" />
-              </Icon>
-            </Button>
+                <Icon
+                  icon="download"
+                  size="small"
+                >
+                  <FormattedMessage id="ui-eholdings.usageConsolidation.summary.actions.export" />
+                </Icon>
+              </Button>
+              {isExportDisabled && (
+                <span className={style['limit-error']}>
+                  <FormattedMessage id="ui-eholdings.usageConsolidation.summary.exportTitles.limit" />
+                </span>
+              )}
+            </div>
           </DropdownMenu>
         )}
       />
@@ -141,6 +156,8 @@ const SummaryTable = ({
 };
 
 SummaryTable.defaultProps = {
+  isExportDisabled: false,
+  onExportTitles: () => {},
   onViewTitles: () => {},
 };